+++ packages/scheduler/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change Log - @lage-run/scheduler
 
-<!-- This log was last generated on Thu, 17 Oct 2024 20:33:04 GMT and should not be manually modified. -->
+<!-- This log was last generated on Mon, 21 Oct 2024 22:18:54 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 1.3.5
+
+Mon, 21 Oct 2024 22:18:54 GMT
+
+### Patches
+
+- Bump @lage-run/cache to v1.3.6
+- Bump @lage-run/config to v0.4.5
+- Bump @lage-run/hasher to v1.6.2
+- Bump @lage-run/runners to v1.0.5
+- Bump @lage-run/target-graph to v0.9.2
+- Bump @lage-run/scheduler-types to v0.3.18
 
 ## 1.3.4